Day #2
	Over the past two days I have been working on, and collaborating with
John, to develop and make the code to read text documents. After a lot
of looking up online of possible ways to do it and multiple trial and
errors we finally figured out how to do it. For my code I needed to
make it so that the loop would go through the whole document, which is
a word list, without stopping while also putting the words into an
array. For this I used an ArrayList and a while loop. I successfully
was able to do this. I then used math.random and the size of the array
to generate a random number in the array and returned the word at that
index. I then put that code into a getWord method in my Dictionary
Class. I also made a get method that gets the whole array. Also for a
proof of concept I have a loop that I commented out that prints out
every word in the array just to make sure it worked (note to self it
crashed java).
	By doing this I have set up the framework for my program so going
forward from here I think I should know how to do everything.