Overhaul
I hadn't been able to commit because I had to redo somethings in my
code. For the most part my code was working but I need to make it do
what I needed it to do. Some of the things in my code I'll remove late
but were just there for testing purposes. I don't think it works
exactly the way that the instructions wanted it to work, but I think I
can fix that. The game is also (in a way) playable. But as it is right
now it is not able to be beaten. I need to implement a code using
math.random so that sometimes it will add letters that the player
guessed.