Beta
Fully functional, improved from class. Fixed the lives and the single
word not being taken out of the list. I haven't test out everything
yet, but the few bugs I know are the mismatch error and somethings
wrong with certain word lengths. As of right now I know that 5 works
fine but for some reason 7 is buggy.